+++ app/Http/Controllers/Manager/SchoolTeacherController.php
@@ -13,12 +13,18 @@ class SchoolTeacherController extends Controller
 {
     public function index(School $school)
     {
+        if($school->id !== auth('api')->user()->schoolID) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         $teachers = $school->teachers;
         return SchoolTeacherResource::collection($teachers);
     }
 
     public function store(AssignTeacherToSchoolRequest $request, School $school)
     {
+        if($school->id !== auth('api')->user()->schoolID) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         $school->teachers()->attach($request->teacherID, ['gradeID' => $request->gradeID]);
 
         return response()->json(['message' => 'Teacher assigned successfully.']);
@@ -26,6 +32,9 @@ class SchoolTeacherController extends Controller
 
     public function update(UpdateSchoolTeacherRequest $request, School $school, Teacher $teacher)
     {
+        if($school->id !== auth('api')->user()->schoolID) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         $school->teachers()->updateExistingPivot($teacher->id, ['gradeID' => $request->gradeID]);
 
         return response()->json(['message' => 'Teacher assignment updated successfully.']);
@@ -33,6 +42,10 @@ class SchoolTeacherController extends Controller
 
     public function destroy(School $school, Teacher $teacher)
     {
+        if($school->id !== auth('api')->user()->schoolID) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $school->teachers()->detach($teacher->id);
 
         return response()->json(null, 204);
